- Renames of adapter, updating import statemets for new package
