+++ dbt/adapters/spark_livy/__version__.py
@@ -1,1 +1,1 @@
-version = "1.2.0a1"
+version = "1.1.0"
