+++ dbt/include/spark_livy/dbt_project.yml
@@ -1,4 +1,4 @@
-name: dbt_spark
+name: dbt_spark_livy
 version: 1.0
 config-version: 2
