+++ tests/unit/test_column.py
@@ -1,6 +1,6 @@
 import unittest
 
-from dbt.adapters.spark import SparkColumn
+from dbt.adapters.spark_livy import SparkColumn
 
 
 class TestSparkColumn(unittest.TestCase):
